+++ config/lib_all.rs
@@ -1,0 +1,8 @@
+// In sync with ../Cargo.toml::lib_all
+// Used from: ../build/features.rs::LIBRARY.features & ../tools/check.rs::LIB_ALL.
+//
+// NOTE: make sure to recompile `../tools/check.rs` after modifying this file.
+// 0 optional dependencie:
+[
+    // "lib_num",
+]